requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==5.2.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.14
wrapt==1.14.1
zope.interface==5.4.0
